+++ lib/support_table_data/documentation.rb
@@ -2,8 +2,10 @@
 
 module SupportTableData
   module Documentation
+    autoload :TypeInference, File.expand_path("documentation/type_inference", __dir__)
+    autoload :SourceFile, File.expand_path("documentation/source_file", __dir__)
+    autoload :YardDoc, File.expand_path("documentation/yard_doc", __dir__)
+    autoload :RbsDoc, File.expand_path("documentation/rbs_doc", __dir__)
+    autoload :RbsFile, File.expand_path("documentation/rbs_file", __dir__)
   end
 end
-
-require_relative "documentation/source_file"
-require_relative "documentation/yard_doc"